requirements.txt
asgiref==3.7.2
bleach==6.0.0
colorama==0.4.6
Django==4.2.5
django-ckeditor==6.7.0     
django-js-asset==2.1.0     
jsmin==3.0.1
Pillow==10.0.1
pywin32==306
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
webencodings==0.5.1